Refactor and finalize project for submission ready version (#13)
* Add navigation link to layout

* Rename and devlop new ViewModel for Statement

* Change namespce using

* Developbuilder for BankStatement

* Refactor view for account current balance
