+++ Views/Shared/_Layout.cshtml
@@ -32,6 +32,9 @@
                             <li class="nav-item">
                                 <a class="nav-link text-dark" asp-controller="BillPay" asp-action="Index">Scheduled Payments</a>
                             </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="BillPay" asp-action="Create">Pay Bills</a>
+                            </li>
                         </ul>
                     </div>
                     <div style="float: right">
